Add files via upload
We introduce the concept of memory-hard proof-of-work algorithms and argue that in order for proof-of-work based systems to be
secure against attacks using custom hardware, they should be constructed from memory-hard functions. Proof-of-work is an economic
measure that is designed to be expensive to find but cheap to verify. Existing solutions used for proof-of-work are either trivial to parallelize or too slow to verify when tuned to use gigabytes of memory. We present a new family of proof-of-work algorithms that can be validated in less than a millisecond but can demand any amount of memory for finding solutions in a practical manner.